Add logging to controllers
A Logger instance is now passed to all controllers and used to log potential errors when rendering Twig templates. In this commit, the BaseController is modified to accept a Logger object and use it for error handling in the render method. Furthermore, the HomeController and ReportController constructors have been updated to receive a Logger object.